requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
greenlet==3.1.1
idna==3.10
pika==1.3.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
